Add files via upload
This section goes into the bet system for the game. It is the functions for the bet system. Based on the set amount you give the program, your bet will be added and subtracted to the total you had prior to betting. This gives the player incentive to keep playing and raise their credit count.